Add files via upload
The ACRelay.py script is a Python-based chat relay system that acts as an intermediary to facilitate communication between clients and other relay servers. The script supports the creation of chat rooms and allows for seamless relay of messages across connected clients and relays. It also includes functionality to generate additional relay scripts or chat room GUI scripts dynamically